% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle.R
\name{circle_area}
\alias{circle_area}
\title{Calculate the area of a circle}
\usage{
circle_area(radius)
}
\arguments{
\item{radius}{The radius of the circle.}
}
\value{
Returns the area of the circle.
}
\description{
This function calculates the area of a circle given its radius.
}
\details{
The formula used is: area = π * radius^2
}
\examples{
circle_area(5)
}
