% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle.R
\name{circle_perimeter}
\alias{circle_perimeter}
\title{Calculate the perimeter of a circle}
\usage{
circle_perimeter(radius)
}
\arguments{
\item{radius}{The radius of the circle.}
}
\value{
Returns the perimeter of the circle.
}
\description{
This function calculates the perimeter of a circle given its radius.
}
\details{
The formula used is: perimeter = 2 * π * radius
}
\examples{
circle_perimeter(5)
}
