% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pol2cart.R
\name{pol2cart}
\alias{pol2cart}
\title{Transform polar or cylindrical coordinates to Cartesian coordinates.}
\usage{
pol2cart(theta, r = NULL, z = NULL)
}
\arguments{
\item{theta}{describes the angle relative to the positive x-axis.}

\item{r}{is the distance to the z-axis (0, 0, z).}

\item{z}{(optional) is the z-coordinate}
}
\value{
a matrix \code{C} where each row represents one Cartesian
  coordinate (\code{x}, \code{y} (, \code{z})).
}
\description{
The inputs \code{theta}, \code{r}, (and \code{z}) must be the same shape, or
scalar.  If called with a single matrix argument then each row of \code{P}
represents the polar/(cylindrical) coordinate (\code{theta}, \code{r}
(, \code{z})).
}
\seealso{
\code{\link{cart2pol}}, \code{\link{sph2cart}},
  \code{\link{cart2sph}}
}
\author{
Kai Habel

David Sterratt
}
