% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangle.R
\name{rectangle_area}
\alias{rectangle_area}
\title{Calculate the area of a rectangle}
\usage{
rectangle_area(length, width)
}
\arguments{
\item{length}{The length of the rectangle.}

\item{width}{The width of the rectangle.}
}
\value{
Returns the area of the rectangle.
}
\description{
This function calculates the area of a rectangle given its length and width.
}
\details{
The formula used is: area = length * width
}
\examples{
rectangle_area(5, 3)
}
