% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangle.R
\name{rectangle_perimeter}
\alias{rectangle_perimeter}
\title{Calculate the perimeter of a rectangle}
\usage{
rectangle_perimeter(length, width)
}
\arguments{
\item{length}{The length of the rectangle.}

\item{width}{The width of the rectangle.}
}
\value{
Returns the perimeter of the rectangle.
}
\description{
This function calculates the perimeter of a rectangle given its length and width.
}
\details{
The formula used is: perimeter = 2 * (length + width)
}
\examples{
rectangle_perimeter(5, 3)
}
