% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convhulln.R
\name{to.mesh3d}
\alias{to.mesh3d}
\title{Convert convhulln object to RGL mesh}
\usage{
to.mesh3d(x, ...)
}
\arguments{
\item{x}{\code{\link{convhulln}} object}

\item{...}{Arguments to \code{\link[rgl]{qmesh3d}} or
\code{\link[rgl]{tmesh3d}}}
}
\value{
\code{\link[rgl]{mesh3d}} object, which can be displayed
  in RGL with \code{\link[rgl]{dot3d}}, \code{\link[rgl]{wire3d}}
  or \code{\link[rgl]{shade3d}}
}
\description{
Convert convhulln object to RGL mesh
}
\seealso{
\code{\link[rgl]{as.mesh3d}}
}
