% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.R
\name{triangle_perimeter}
\alias{triangle_perimeter}
\title{Calculate the perimeter of a triangle}
\usage{
triangle_perimeter(a, b, c)
}
\arguments{
\item{a}{The length of the first side of the triangle.}

\item{b}{The length of the second side of the triangle.}

\item{c}{The length of the third side of the triangle.}
}
\value{
Returns the perimeter of the triangle.
}
\description{
This function calculates the perimeter of a triangle given the lengths of its sides.
}
\details{
The formula used is: perimeter = a + b + c
}
\examples{
triangle_perimeter(3, 4, 5)
}
