% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbox.R
\name{rbox}
\alias{rbox}
\title{Generate various point distributions}
\usage{
rbox(D = 2, B = 0.5)
}
\arguments{
\item{D}{number of dimensions of hypercube}

\item{B}{bounding box coordinate - faces will be \code{-B} and \code{B} from origin}
}
\value{
Matrix of points
}
\description{
Default is corners of a hypercube.
}
\author{
David Sterratt
}

